Move reusable dashboard models
