+++ pir/pir-impl/src/main/java/com/duckduckgo/pir/impl/dashboard/state/PirDashboardModels.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2025 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.pir.impl.dashboard.state
+
+import com.duckduckgo.pir.impl.models.ExtractedProfile
+
+data class DashboardBroker(
+    val name: String,
+    val url: String,
+    val parentUrl: String? = null,
+    val optOutUrl: String? = null,
+)
+
+/**
+ * This class represents a valid [ExtractedProfile] that should be displayed in the dashboard.
+ */
+data class DashboardExtractedProfileResult(
+    val extractedProfile: ExtractedProfile,
+    val broker: DashboardBroker,
+    val optOutSubmittedDateInMillis: Long? = null,
+    val optOutRemovedDateInMillis: Long? = null,
+    val estimatedRemovalDateInMillis: Long? = null,
+    val hasMatchingRecordOnParentBroker: Boolean = false,
+)
